+++ middleware/contacts/index.js
@@ -1,0 +1,17 @@
+const getAllContacts = require("./getAll.js");
+const validateContact = require("./validation.js");
+const saveAllContact = require("./saveContact.js");
+const getALLbyId = require("./getById.js");
+const updateFavoriteStatus = require("./updateFavoriteStatus.js");
+const deleteContact = require("./deleteContact.js");
+const updateAllContact = require("./updateContact.js");
+
+module.exports = {
+    getAllContacts,
+    validateContact,
+    saveAllContact,
+    getALLbyId,
+    updateFavoriteStatus,
+    deleteContact,
+    updateAllContact,
+}
