+++ middleware/contacts/validation.js
@@ -1,0 +1,18 @@
+const Joi = require("joi");
+
+const validateContact = (req, res, next) => {
+  const contactPostSchema = Joi.object({
+    name: Joi.string().required(),
+    email: Joi.string().email().required(),
+    phone: Joi.string().required(),
+    favorite: Joi.boolean(),
+  });
+
+  const { error } = contactPostSchema.validate(req.body);
+  if (error) {
+    return res.status(400).json({ message: error.details[0].message });
+  }
+  next();
+};
+
+module.exports = validateContact;
