+++ middleware/index.js
@@ -1,0 +1,5 @@
+const contacts = require("./contacts/index.js");
+
+module.exports = {
+    contacts,
+}
